+++ packages/platform-browser/src/browser/transfer_state_module.ts
